qi/core/tests/unit/data/models/base/tick.test.ts
/**
 * @fileoverview
 * @module tick.test.ts
 *
 * @author zhifengzhang-sz
 * @created 2024-12-11
 * @modified 2024-12-11
 */

// tick.test.ts
import { describe, expect, it } from "vitest";
import { Tick } from "@qi/core/data/models/base/tick";

describe("Tick", () => {
  const now = Date.now();

  it("should create a valid Tick object", () => {
    const tick: Tick = {
      timestamp: now,
      price: 100.5,
      quantity: 10,
      side: "buy",
      exchange: "NYSE",
      symbol: "AAPL",
    };

    expect(tick.timestamp).toBe(now);
    expect(tick.price).toBe(100.5);
    expect(tick.quantity).toBe(10);
    expect(tick.side).toBe("buy");
    expect(tick.exchange).toBe("NYSE");
    expect(tick.symbol).toBe("AAPL");
  });

  it("should accept sell side trades", () => {
    const tick: Tick = {
      timestamp: now,
      price: 100.5,
      quantity: 10,
      side: "sell",
      exchange: "NYSE",
      symbol: "AAPL",
    };
    expect(tick.side).toBe("sell");
  });

  it("should accept unknown side trades", () => {
    const tick: Tick = {
      timestamp: now,
      price: 100.5,
      quantity: 10,
      side: "unknown",
      exchange: "NYSE",
      symbol: "AAPL",
    };
    expect(tick.side).toBe("unknown");
  });

  it("should handle decimal quantities and prices", () => {
    const tick: Tick = {
      timestamp: now,
      price: 100.12345678,
      quantity: 0.12345678,
      side: "buy",
      exchange: "NYSE",
      symbol: "AAPL",
    };
    expect(tick.price).toBe(100.12345678);
    expect(tick.quantity).toBe(0.12345678);
  });

  describe("type checking", () => {
    // Note: The following type checks are commented out because they should fail compilation.
    // They serve as documentation for the type system's behavior.
    it("should not allow invalid side values", () => {
      // TypeScript should prevent this:
      // const invalidTick: Tick = {
      //   timestamp: now,
      //   price: 100.5,
      //   quantity: 10,
      //   side: "invalid", // Type error: must be "buy" | "sell" | "unknown"
      //   exchange: "NYSE",
      //   symbol: "AAPL",
      // };
      expect(true).toBe(true); // Placeholder assertion
    });

    it("should require all mandatory fields", () => {
      // TypeScript should prevent this:
      // const incompleteTick: Tick = {
      //   timestamp: now,
      //   price: 100.5,
      //   quantity: 10,
      //   // side: "buy",  // Type error: missing required property 'side'
      //   exchange: "NYSE",
      //   symbol: "AAPL",
      // };
      expect(true).toBe(true); // Placeholder assertion
    });
  });
});
